requirements.txt
fastapi==0.110.0
pydantic==2.6.4
aiogram==3.6.0
aiohttp==3.9.3
uvicorn==0.29.0
python-dotenv==1.0.1